+++ inst/ShinyItemAnalysis/ui.R
@@ -127,7 +127,7 @@ ui <- tagList(
     ),
     tags$script(
       HTML('
-      Shiny.addCustomMessageHandler("sessinf",
+      Shiny.addCustomMessageHandler("send_to_console",
         function(x) {
           console.log(x);
         }
@@ -157,9 +157,9 @@ ui <- tagList(
            <div class = "panel-footer", style = "opacity: 1.00; z-index: 1000;">
               <p style = "margin:8px 0 0 0;">
                 <div class = "footer-title"> <img src = "sia_logo.svg" style="width: 57px; margin-right: 10px;"> ShinyItemAnalysis </div>
-                <div class = "footer-subtitle"> Test and item analysis via Shiny | Version 1.5.0 </div>
+                <div class = "footer-subtitle"> Test and item analysis via Shiny | Version 1.5.1 </div>
                 <span style = "float:right">
-                  <a href = "http://shinyitemanalysis.org/" id = "tooltipweb" target="_blank"> <img src = "footer_web_icon.png", class = "footer-icons"> </a>
+                  <a href = "https://www.shinyitemanalysis.org/" id = "tooltipweb" target="_blank"> <img src = "footer_web_icon.png", class = "footer-icons"> </a>
                   <a href = "https://github.com/patriciamar/ShinyItemAnalysis/" id = "tooltipgithub" target="_blank"> <img src = "footer_github_icon.png", class = "footer-icons"> </a>
                   <a href = "https://CRAN.R-project.org/package=ShinyItemAnalysis/" id = "tooltipcran" target="_blank"> <img src = "footer_cran_icon.png", class = "footer-icons"> </a>
                 </span>